[wip] Migrate and refurbish search related endpoints to api/search.py
