+++ src/api/__init__.py
@@ -1,6 +1,8 @@
 from fastapi import APIRouter
 
 from src.api.auth import router as auth_router
+from src.api.search import router as search_router
 
 api = APIRouter()
 api.include_router(auth_router, prefix="/auth", tags=["auth"])
+api.include_router(search_router, prefix="/search", tags=["search"])
