+++ templates/search.html
@@ -1,6 +1,6 @@
 <div class='main_content'>
     <div class='pt-2 mt-20 items-center justify-center px-2 h-full'>
-        <form method="GET" hx-target="next .search_table" hx-get="/htmx_search" hx-indicator="#head-spinner">
+        <form method="GET" hx-target="next .search_table" hx-get="/search/results" hx-indicator="#head-spinner">
 
             <div class="flex justify-center items-center space-x-2 pt-2">
                 <input type="checkbox" id="e_resources" name="e_resources" value=True
